requirements.txt
elasticsearch
langchain
lingua
numpy
pandas
pydantic
python-dotenv
requests
sentence-transformers
spacy
